Refactor StrategyGroup and utilities into separate files
Split StrategyGroup and its supporting utilities into StrategyGroup.cs and StrategyGroup.Utilities.cs for better modularity.
Moved BackPair<T> and StrategyGroupUtilities (with TryGetBackPair and AreAligned) to the new utilities file, enhancing documentation and adding an offset parameter to TryGetBackPair.
Updated StrategyGroup classes to use primary constructors and improved XML docs throughout.
This refactor improves code clarity, maintainability, and organization.
